refactor: mover configs e integrações externas para infrastructure
