+++ src/main/java/com/sensedia/sample/consents/application/service/impl/ConsentServiceImpl.java
@@ -1,6 +1,6 @@
 package com.sensedia.sample.consents.application.service.impl;
 
-import com.sensedia.sample.consents.client.GitHubClient;
+import com.sensedia.sample.consents.infra.client.GitHubClient;
 import com.sensedia.sample.consents.domain.model.Consent;
 import com.sensedia.sample.consents.domain.model.ConsentHistory;
 import com.sensedia.sample.consents.rest.dto.ConsentRequestDTO;
